Revisit Time Reference Ground Tracks and Dates

This function shows the information of the 'available_RGTs' function and additionally it returns the .zip (kmz files) of all laser tracks over each 91-day repeat period (revisit time). Note that the locations and times are estimates, but should be correct to within a few minutes in time and better than 100m in the predicted locations.

revisit_time_RGTs(RGT_cycle = NULL, complete_date_sequence = FALSE)

Arguments

RGT_cycle

either NULL or a character string specifying a single RGT (Reference Ground Track) as determined by the output of the 'available_RGTs(only_cycle_names = TRUE)' function. If NULL then all available Data will be returned.

complete_date_sequence

a boolean. If TRUE a complete sequence of Dates will be returned, otherwise only the 'minimum' and 'maximum' Dates.

Value

a list object with the available orbit files, dates and date sequence lengths

Details

ICESat-2 was in safe-hold from June 26 through July 9, 2019. ATLAS was off during this time, so data was not collected or pointed to the reference ground track.

References

https://icesat-2.gsfc.nasa.gov/science/specs

Examples

require(IceSat2R)

#.......................................................
# receive all orbit files, dates and length of sequences
#.......................................................

rev_all = revisit_time_RGTs(RGT_cycle = NULL, complete_date_sequence = TRUE)
rev_all
#> $orbit_files
#> $orbit_files$RGT_cycle_1
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle1_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_2
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle2_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_3
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle3_date_time_0.zip"
#> 
#> $orbit_files$RGT_cycle_4
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle4_date_time_rev2_0.zip"
#> 
#> $orbit_files$RGT_cycle_5
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle5_date_time_rev3.zip"
#> 
#> $orbit_files$RGT_cycle_6
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle6_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_7
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle7_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_8
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle8_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_9
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle9_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_10
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle10_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_11
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle11_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_12
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2RGTscycle12datetime.zip"
#> 
#> $orbit_files$RGT_cycle_13
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle13_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_14
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle14_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_15
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle15_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_16
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle16_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_17
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle17_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_18
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle18_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_19
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle19_date_time.zip"
#> 
#> $orbit_files$RGT_cycle_20
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle20_date_time_provisional.zip"
#> 
#> 
#> $orbit_dates
#> $orbit_dates$RGT_cycle_1
#>  [1] "2018-10-13" "2018-10-14" "2018-10-15" "2018-10-16" "2018-10-17"
#>  [6] "2018-10-18" "2018-10-19" "2018-10-20" "2018-10-21" "2018-10-22"
#> [11] "2018-10-23" "2018-10-24" "2018-10-25" "2018-10-26" "2018-10-27"
#> [16] "2018-10-28" "2018-10-29" "2018-10-30" "2018-10-31" "2018-11-01"
#> [21] "2018-11-02" "2018-11-03" "2018-11-04" "2018-11-05" "2018-11-06"
#> [26] "2018-11-07" "2018-11-08" "2018-11-09" "2018-11-10" "2018-11-11"
#> [31] "2018-11-12" "2018-11-13" "2018-11-14" "2018-11-15" "2018-11-16"
#> [36] "2018-11-17" "2018-11-18" "2018-11-19" "2018-11-20" "2018-11-21"
#> [41] "2018-11-22" "2018-11-23" "2018-11-24" "2018-11-25" "2018-11-26"
#> [46] "2018-11-27" "2018-11-28" "2018-11-29" "2018-11-30" "2018-12-01"
#> [51] "2018-12-02" "2018-12-03" "2018-12-04" "2018-12-05" "2018-12-06"
#> [56] "2018-12-07" "2018-12-08" "2018-12-09" "2018-12-10" "2018-12-11"
#> [61] "2018-12-12" "2018-12-13" "2018-12-14" "2018-12-15" "2018-12-16"
#> [66] "2018-12-17" "2018-12-18" "2018-12-19" "2018-12-20" "2018-12-21"
#> [71] "2018-12-22" "2018-12-23" "2018-12-24" "2018-12-25" "2018-12-26"
#> [76] "2018-12-27" "2018-12-28"
#> 
#> $orbit_dates$RGT_cycle_2
#>  [1] "2018-12-28" "2018-12-29" "2018-12-30" "2018-12-31" "2019-01-01"
#>  [6] "2019-01-02" "2019-01-03" "2019-01-04" "2019-01-05" "2019-01-06"
#> [11] "2019-01-07" "2019-01-08" "2019-01-09" "2019-01-10" "2019-01-11"
#> [16] "2019-01-12" "2019-01-13" "2019-01-14" "2019-01-15" "2019-01-16"
#> [21] "2019-01-17" "2019-01-18" "2019-01-19" "2019-01-20" "2019-01-21"
#> [26] "2019-01-22" "2019-01-23" "2019-01-24" "2019-01-25" "2019-01-26"
#> [31] "2019-01-27" "2019-01-28" "2019-01-29" "2019-01-30" "2019-01-31"
#> [36] "2019-02-01" "2019-02-02" "2019-02-03" "2019-02-04" "2019-02-05"
#> [41] "2019-02-06" "2019-02-07" "2019-02-08" "2019-02-09" "2019-02-10"
#> [46] "2019-02-11" "2019-02-12" "2019-02-13" "2019-02-14" "2019-02-15"
#> [51] "2019-02-16" "2019-02-17" "2019-02-18" "2019-02-19" "2019-02-20"
#> [56] "2019-02-21" "2019-02-22" "2019-02-23" "2019-02-24" "2019-02-25"
#> [61] "2019-02-26" "2019-02-27" "2019-02-28" "2019-03-01" "2019-03-02"
#> [66] "2019-03-03" "2019-03-04" "2019-03-05" "2019-03-06" "2019-03-07"
#> [71] "2019-03-08" "2019-03-09" "2019-03-10" "2019-03-11" "2019-03-12"
#> [76] "2019-03-13" "2019-03-14" "2019-03-15" "2019-03-16" "2019-03-17"
#> [81] "2019-03-18" "2019-03-19" "2019-03-20" "2019-03-21" "2019-03-22"
#> [86] "2019-03-23" "2019-03-24" "2019-03-25" "2019-03-26" "2019-03-27"
#> [91] "2019-03-28" "2019-03-29"
#> 
#> $orbit_dates$RGT_cycle_3
#>  [1] "2019-03-29" "2019-03-30" "2019-03-31" "2019-04-01" "2019-04-02"
#>  [6] "2019-04-03" "2019-04-04" "2019-04-05" "2019-04-06" "2019-04-07"
#> [11] "2019-04-08" "2019-04-09" "2019-04-10" "2019-04-11" "2019-04-12"
#> [16] "2019-04-13" "2019-04-14" "2019-04-15" "2019-04-16" "2019-04-17"
#> [21] "2019-04-18" "2019-04-19" "2019-04-20" "2019-04-21" "2019-04-22"
#> [26] "2019-04-23" "2019-04-24" "2019-04-25" "2019-04-26" "2019-04-27"
#> [31] "2019-04-28" "2019-04-29" "2019-04-30" "2019-05-01" "2019-05-02"
#> [36] "2019-05-03" "2019-05-04" "2019-05-05" "2019-05-06" "2019-05-07"
#> [41] "2019-05-08" "2019-05-09" "2019-05-10" "2019-05-11" "2019-05-12"
#> [46] "2019-05-13" "2019-05-14" "2019-05-15" "2019-05-16" "2019-05-17"
#> [51] "2019-05-18" "2019-05-19" "2019-05-20" "2019-05-21" "2019-05-22"
#> [56] "2019-05-23" "2019-05-24" "2019-05-25" "2019-05-26" "2019-05-27"
#> [61] "2019-05-28" "2019-05-29" "2019-05-30" "2019-05-31" "2019-06-01"
#> [66] "2019-06-02" "2019-06-03" "2019-06-04" "2019-06-05" "2019-06-06"
#> [71] "2019-06-07" "2019-06-08" "2019-06-09" "2019-06-10" "2019-06-11"
#> [76] "2019-06-12" "2019-06-13" "2019-06-14" "2019-06-15" "2019-06-16"
#> [81] "2019-06-17" "2019-06-18" "2019-06-19" "2019-06-20" "2019-06-21"
#> [86] "2019-06-22" "2019-06-23" "2019-06-24" "2019-06-25" "2019-06-26"
#> [91] "2019-06-27" "2019-06-28"
#> 
#> $orbit_dates$RGT_cycle_4
#>  [1] "2019-07-09" "2019-07-10" "2019-07-11" "2019-07-12" "2019-07-13"
#>  [6] "2019-07-14" "2019-07-15" "2019-07-16" "2019-07-17" "2019-07-18"
#> [11] "2019-07-19" "2019-07-20" "2019-07-21" "2019-07-22" "2019-07-23"
#> [16] "2019-07-24" "2019-07-25" "2019-07-26" "2019-07-27" "2019-07-28"
#> [21] "2019-07-29" "2019-07-30" "2019-07-31" "2019-08-01" "2019-08-02"
#> [26] "2019-08-03" "2019-08-04" "2019-08-05" "2019-08-06" "2019-08-07"
#> [31] "2019-08-08" "2019-08-09" "2019-08-10" "2019-08-11" "2019-08-12"
#> [36] "2019-08-13" "2019-08-14" "2019-08-15" "2019-08-16" "2019-08-17"
#> [41] "2019-08-18" "2019-08-19" "2019-08-20" "2019-08-21" "2019-08-22"
#> [46] "2019-08-23" "2019-08-24" "2019-08-25" "2019-08-26" "2019-08-27"
#> [51] "2019-08-28" "2019-08-29" "2019-08-30" "2019-08-31" "2019-09-01"
#> [56] "2019-09-02" "2019-09-03" "2019-09-04" "2019-09-05" "2019-09-06"
#> [61] "2019-09-07" "2019-09-08" "2019-09-09" "2019-09-10" "2019-09-11"
#> [66] "2019-09-12" "2019-09-13" "2019-09-14" "2019-09-15" "2019-09-16"
#> [71] "2019-09-17" "2019-09-18" "2019-09-19" "2019-09-20" "2019-09-21"
#> [76] "2019-09-22" "2019-09-23" "2019-09-24" "2019-09-25" "2019-09-26"
#> 
#> $orbit_dates$RGT_cycle_5
#>  [1] "2019-09-26" "2019-09-27" "2019-09-28" "2019-09-29" "2019-09-30"
#>  [6] "2019-10-01" "2019-10-02" "2019-10-03" "2019-10-04" "2019-10-05"
#> [11] "2019-10-06" "2019-10-07" "2019-10-08" "2019-10-09" "2019-10-10"
#> [16] "2019-10-11" "2019-10-12" "2019-10-13" "2019-10-14" "2019-10-15"
#> [21] "2019-10-16" "2019-10-17" "2019-10-18" "2019-10-19" "2019-10-20"
#> [26] "2019-10-21" "2019-10-22" "2019-10-23" "2019-10-24" "2019-10-25"
#> [31] "2019-10-26" "2019-10-27" "2019-10-28" "2019-10-29" "2019-10-30"
#> [36] "2019-10-31" "2019-11-01" "2019-11-02" "2019-11-03" "2019-11-04"
#> [41] "2019-11-05" "2019-11-06" "2019-11-07" "2019-11-08" "2019-11-09"
#> [46] "2019-11-10" "2019-11-11" "2019-11-12" "2019-11-13" "2019-11-14"
#> [51] "2019-11-15" "2019-11-16" "2019-11-17" "2019-11-18" "2019-11-19"
#> [56] "2019-11-20" "2019-11-21" "2019-11-22" "2019-11-23" "2019-11-24"
#> [61] "2019-11-25" "2019-11-26" "2019-11-27" "2019-11-28" "2019-11-29"
#> [66] "2019-11-30" "2019-12-01" "2019-12-02" "2019-12-03" "2019-12-04"
#> [71] "2019-12-05" "2019-12-06" "2019-12-07" "2019-12-08" "2019-12-09"
#> [76] "2019-12-10" "2019-12-11" "2019-12-12" "2019-12-13" "2019-12-14"
#> [81] "2019-12-15" "2019-12-16" "2019-12-17" "2019-12-18" "2019-12-19"
#> [86] "2019-12-20" "2019-12-21" "2019-12-22" "2019-12-23" "2019-12-24"
#> [91] "2019-12-25" "2019-12-26"
#> 
#> $orbit_dates$RGT_cycle_6
#>  [1] "2019-12-26" "2019-12-27" "2019-12-28" "2019-12-29" "2019-12-30"
#>  [6] "2019-12-31" "2020-01-01" "2020-01-02" "2020-01-03" "2020-01-04"
#> [11] "2020-01-05" "2020-01-06" "2020-01-07" "2020-01-08" "2020-01-09"
#> [16] "2020-01-10" "2020-01-11" "2020-01-12" "2020-01-13" "2020-01-14"
#> [21] "2020-01-15" "2020-01-16" "2020-01-17" "2020-01-18" "2020-01-19"
#> [26] "2020-01-20" "2020-01-21" "2020-01-22" "2020-01-23" "2020-01-24"
#> [31] "2020-01-25" "2020-01-26" "2020-01-27" "2020-01-28" "2020-01-29"
#> [36] "2020-01-30" "2020-01-31" "2020-02-01" "2020-02-02" "2020-02-03"
#> [41] "2020-02-04" "2020-02-05" "2020-02-06" "2020-02-07" "2020-02-08"
#> [46] "2020-02-09" "2020-02-10" "2020-02-11" "2020-02-12" "2020-02-13"
#> [51] "2020-02-14" "2020-02-15" "2020-02-16" "2020-02-17" "2020-02-18"
#> [56] "2020-02-19" "2020-02-20" "2020-02-21" "2020-02-22" "2020-02-23"
#> [61] "2020-02-24" "2020-02-25" "2020-02-26" "2020-02-27" "2020-02-28"
#> [66] "2020-02-29" "2020-03-01" "2020-03-02" "2020-03-03" "2020-03-04"
#> [71] "2020-03-05" "2020-03-06" "2020-03-07" "2020-03-08" "2020-03-09"
#> [76] "2020-03-10" "2020-03-11" "2020-03-12" "2020-03-13" "2020-03-14"
#> [81] "2020-03-15" "2020-03-16" "2020-03-17" "2020-03-18" "2020-03-19"
#> [86] "2020-03-20" "2020-03-21" "2020-03-22" "2020-03-23" "2020-03-24"
#> [91] "2020-03-25" "2020-03-26"
#> 
#> $orbit_dates$RGT_cycle_7
#>  [1] "2020-03-26" "2020-03-27" "2020-03-28" "2020-03-29" "2020-03-30"
#>  [6] "2020-03-31" "2020-04-01" "2020-04-02" "2020-04-03" "2020-04-04"
#> [11] "2020-04-05" "2020-04-06" "2020-04-07" "2020-04-08" "2020-04-09"
#> [16] "2020-04-10" "2020-04-11" "2020-04-12" "2020-04-13" "2020-04-14"
#> [21] "2020-04-15" "2020-04-16" "2020-04-17" "2020-04-18" "2020-04-19"
#> [26] "2020-04-20" "2020-04-21" "2020-04-22" "2020-04-23" "2020-04-24"
#> [31] "2020-04-25" "2020-04-26" "2020-04-27" "2020-04-28" "2020-04-29"
#> [36] "2020-04-30" "2020-05-01" "2020-05-02" "2020-05-03" "2020-05-04"
#> [41] "2020-05-05" "2020-05-06" "2020-05-07" "2020-05-08" "2020-05-09"
#> [46] "2020-05-10" "2020-05-11" "2020-05-12" "2020-05-13" "2020-05-14"
#> [51] "2020-05-15" "2020-05-16" "2020-05-17" "2020-05-18" "2020-05-19"
#> [56] "2020-05-20" "2020-05-21" "2020-05-22" "2020-05-23" "2020-05-24"
#> [61] "2020-05-25" "2020-05-26" "2020-05-27" "2020-05-28" "2020-05-29"
#> [66] "2020-05-30" "2020-05-31" "2020-06-01" "2020-06-02" "2020-06-03"
#> [71] "2020-06-04" "2020-06-05" "2020-06-06" "2020-06-07" "2020-06-08"
#> [76] "2020-06-09" "2020-06-10" "2020-06-11" "2020-06-12" "2020-06-13"
#> [81] "2020-06-14" "2020-06-15" "2020-06-16" "2020-06-17" "2020-06-18"
#> [86] "2020-06-19" "2020-06-20" "2020-06-21" "2020-06-22" "2020-06-23"
#> [91] "2020-06-24" "2020-06-25"
#> 
#> $orbit_dates$RGT_cycle_8
#>  [1] "2020-06-25" "2020-06-26" "2020-06-27" "2020-06-28" "2020-06-29"
#>  [6] "2020-06-30" "2020-07-01" "2020-07-02" "2020-07-03" "2020-07-04"
#> [11] "2020-07-05" "2020-07-06" "2020-07-07" "2020-07-08" "2020-07-09"
#> [16] "2020-07-10" "2020-07-11" "2020-07-12" "2020-07-13" "2020-07-14"
#> [21] "2020-07-15" "2020-07-16" "2020-07-17" "2020-07-18" "2020-07-19"
#> [26] "2020-07-20" "2020-07-21" "2020-07-22" "2020-07-23" "2020-07-24"
#> [31] "2020-07-25" "2020-07-26" "2020-07-27" "2020-07-28" "2020-07-29"
#> [36] "2020-07-30" "2020-07-31" "2020-08-01" "2020-08-02" "2020-08-03"
#> [41] "2020-08-04" "2020-08-05" "2020-08-06" "2020-08-07" "2020-08-08"
#> [46] "2020-08-09" "2020-08-10" "2020-08-11" "2020-08-12" "2020-08-13"
#> [51] "2020-08-14" "2020-08-15" "2020-08-16" "2020-08-17" "2020-08-18"
#> [56] "2020-08-19" "2020-08-20" "2020-08-21" "2020-08-22" "2020-08-23"
#> [61] "2020-08-24" "2020-08-25" "2020-08-26" "2020-08-27" "2020-08-28"
#> [66] "2020-08-29" "2020-08-30" "2020-08-31" "2020-09-01" "2020-09-02"
#> [71] "2020-09-03" "2020-09-04" "2020-09-05" "2020-09-06" "2020-09-07"
#> [76] "2020-09-08" "2020-09-09" "2020-09-10" "2020-09-11" "2020-09-12"
#> [81] "2020-09-13" "2020-09-14" "2020-09-15" "2020-09-16" "2020-09-17"
#> [86] "2020-09-18" "2020-09-19" "2020-09-20" "2020-09-21" "2020-09-22"
#> [91] "2020-09-23" "2020-09-24"
#> 
#> $orbit_dates$RGT_cycle_9
#>  [1] "2020-09-24" "2020-09-25" "2020-09-26" "2020-09-27" "2020-09-28"
#>  [6] "2020-09-29" "2020-09-30" "2020-10-01" "2020-10-02" "2020-10-03"
#> [11] "2020-10-04" "2020-10-05" "2020-10-06" "2020-10-07" "2020-10-08"
#> [16] "2020-10-09" "2020-10-10" "2020-10-11" "2020-10-12" "2020-10-13"
#> [21] "2020-10-14" "2020-10-15" "2020-10-16" "2020-10-17" "2020-10-18"
#> [26] "2020-10-19" "2020-10-20" "2020-10-21" "2020-10-22" "2020-10-23"
#> [31] "2020-10-24" "2020-10-25" "2020-10-26" "2020-10-27" "2020-10-28"
#> [36] "2020-10-29" "2020-10-30" "2020-10-31" "2020-11-01" "2020-11-02"
#> [41] "2020-11-03" "2020-11-04" "2020-11-05" "2020-11-06" "2020-11-07"
#> [46] "2020-11-08" "2020-11-09" "2020-11-10" "2020-11-11" "2020-11-12"
#> [51] "2020-11-13" "2020-11-14" "2020-11-15" "2020-11-16" "2020-11-17"
#> [56] "2020-11-18" "2020-11-19" "2020-11-20" "2020-11-21" "2020-11-22"
#> [61] "2020-11-23" "2020-11-24" "2020-11-25" "2020-11-26" "2020-11-27"
#> [66] "2020-11-28" "2020-11-29" "2020-11-30" "2020-12-01" "2020-12-02"
#> [71] "2020-12-03" "2020-12-04" "2020-12-05" "2020-12-06" "2020-12-07"
#> [76] "2020-12-08" "2020-12-09" "2020-12-10" "2020-12-11" "2020-12-12"
#> [81] "2020-12-13" "2020-12-14" "2020-12-15" "2020-12-16" "2020-12-17"
#> [86] "2020-12-18" "2020-12-19" "2020-12-20" "2020-12-21" "2020-12-22"
#> [91] "2020-12-23"
#> 
#> $orbit_dates$RGT_cycle_10
#>  [1] "2020-12-24" "2020-12-25" "2020-12-26" "2020-12-27" "2020-12-28"
#>  [6] "2020-12-29" "2020-12-30" "2020-12-31" "2021-01-01" "2021-01-02"
#> [11] "2021-01-03" "2021-01-04" "2021-01-05" "2021-01-06" "2021-01-07"
#> [16] "2021-01-08" "2021-01-09" "2021-01-10" "2021-01-11" "2021-01-12"
#> [21] "2021-01-13" "2021-01-14" "2021-01-15" "2021-01-16" "2021-01-17"
#> [26] "2021-01-18" "2021-01-19" "2021-01-20" "2021-01-21" "2021-01-22"
#> [31] "2021-01-23" "2021-01-24" "2021-01-25" "2021-01-26" "2021-01-27"
#> [36] "2021-01-28" "2021-01-29" "2021-01-30" "2021-01-31" "2021-02-01"
#> [41] "2021-02-02" "2021-02-03" "2021-02-04" "2021-02-05" "2021-02-06"
#> [46] "2021-02-07" "2021-02-08" "2021-02-09" "2021-02-10" "2021-02-11"
#> [51] "2021-02-12" "2021-02-13" "2021-02-14" "2021-02-15" "2021-02-16"
#> [56] "2021-02-17" "2021-02-18" "2021-02-19" "2021-02-20" "2021-02-21"
#> [61] "2021-02-22" "2021-02-23" "2021-02-24" "2021-02-25" "2021-02-26"
#> [66] "2021-02-27" "2021-02-28" "2021-03-01" "2021-03-02" "2021-03-03"
#> [71] "2021-03-04" "2021-03-05" "2021-03-06" "2021-03-07" "2021-03-08"
#> [76] "2021-03-09" "2021-03-10" "2021-03-11" "2021-03-12" "2021-03-13"
#> [81] "2021-03-14" "2021-03-15" "2021-03-16" "2021-03-17" "2021-03-18"
#> [86] "2021-03-19" "2021-03-20" "2021-03-21" "2021-03-22" "2021-03-23"
#> [91] "2021-03-24"
#> 
#> $orbit_dates$RGT_cycle_11
#>  [1] "2021-03-24" "2021-03-25" "2021-03-26" "2021-03-27" "2021-03-28"
#>  [6] "2021-03-29" "2021-03-30" "2021-03-31" "2021-04-01" "2021-04-02"
#> [11] "2021-04-03" "2021-04-04" "2021-04-05" "2021-04-06" "2021-04-07"
#> [16] "2021-04-08" "2021-04-09" "2021-04-10" "2021-04-11" "2021-04-12"
#> [21] "2021-04-13" "2021-04-14" "2021-04-15" "2021-04-16" "2021-04-17"
#> [26] "2021-04-18" "2021-04-19" "2021-04-20" "2021-04-21" "2021-04-22"
#> [31] "2021-04-23" "2021-04-24" "2021-04-25" "2021-04-26" "2021-04-27"
#> [36] "2021-04-28" "2021-04-29" "2021-04-30" "2021-05-01" "2021-05-02"
#> [41] "2021-05-03" "2021-05-04" "2021-05-05" "2021-05-06" "2021-05-07"
#> [46] "2021-05-08" "2021-05-09" "2021-05-10" "2021-05-11" "2021-05-12"
#> [51] "2021-05-13" "2021-05-14" "2021-05-15" "2021-05-16" "2021-05-17"
#> [56] "2021-05-18" "2021-05-19" "2021-05-20" "2021-05-21" "2021-05-22"
#> [61] "2021-05-23" "2021-05-24" "2021-05-25" "2021-05-26" "2021-05-27"
#> [66] "2021-05-28" "2021-05-29" "2021-05-30" "2021-05-31" "2021-06-01"
#> [71] "2021-06-02" "2021-06-03" "2021-06-04" "2021-06-05" "2021-06-06"
#> [76] "2021-06-07" "2021-06-08" "2021-06-09" "2021-06-10" "2021-06-11"
#> [81] "2021-06-12" "2021-06-13" "2021-06-14" "2021-06-15" "2021-06-16"
#> [86] "2021-06-17" "2021-06-18" "2021-06-19" "2021-06-20" "2021-06-21"
#> [91] "2021-06-22" "2021-06-23"
#> 
#> $orbit_dates$RGT_cycle_12
#>  [1] "2021-06-23" "2021-06-24" "2021-06-25" "2021-06-26" "2021-06-27"
#>  [6] "2021-06-28" "2021-06-29" "2021-06-30" "2021-07-01" "2021-07-02"
#> [11] "2021-07-03" "2021-07-04" "2021-07-05" "2021-07-06" "2021-07-07"
#> [16] "2021-07-08" "2021-07-09" "2021-07-10" "2021-07-11" "2021-07-12"
#> [21] "2021-07-13" "2021-07-14" "2021-07-15" "2021-07-16" "2021-07-17"
#> [26] "2021-07-18" "2021-07-19" "2021-07-20" "2021-07-21" "2021-07-22"
#> [31] "2021-07-23" "2021-07-24" "2021-07-25" "2021-07-26" "2021-07-27"
#> [36] "2021-07-28" "2021-07-29" "2021-07-30" "2021-07-31" "2021-08-01"
#> [41] "2021-08-02" "2021-08-03" "2021-08-04" "2021-08-05" "2021-08-06"
#> [46] "2021-08-07" "2021-08-08" "2021-08-09" "2021-08-10" "2021-08-11"
#> [51] "2021-08-12" "2021-08-13" "2021-08-14" "2021-08-15" "2021-08-16"
#> [56] "2021-08-17" "2021-08-18" "2021-08-19" "2021-08-20" "2021-08-21"
#> [61] "2021-08-22" "2021-08-23" "2021-08-24" "2021-08-25" "2021-08-26"
#> [66] "2021-08-27" "2021-08-28" "2021-08-29" "2021-08-30" "2021-08-31"
#> [71] "2021-09-01" "2021-09-02" "2021-09-03" "2021-09-04" "2021-09-05"
#> [76] "2021-09-06" "2021-09-07" "2021-09-08" "2021-09-09" "2021-09-10"
#> [81] "2021-09-11" "2021-09-12" "2021-09-13" "2021-09-14" "2021-09-15"
#> [86] "2021-09-16" "2021-09-17" "2021-09-18" "2021-09-19" "2021-09-20"
#> [91] "2021-09-21" "2021-09-22"
#> 
#> $orbit_dates$RGT_cycle_13
#>  [1] "2021-09-22" "2021-09-23" "2021-09-24" "2021-09-25" "2021-09-26"
#>  [6] "2021-09-27" "2021-09-28" "2021-09-29" "2021-09-30" "2021-10-01"
#> [11] "2021-10-02" "2021-10-03" "2021-10-04" "2021-10-05" "2021-10-06"
#> [16] "2021-10-07" "2021-10-08" "2021-10-09" "2021-10-10" "2021-10-11"
#> [21] "2021-10-12" "2021-10-13" "2021-10-14" "2021-10-15" "2021-10-16"
#> [26] "2021-10-17" "2021-10-18" "2021-10-19" "2021-10-20" "2021-10-21"
#> [31] "2021-10-22" "2021-10-23" "2021-10-24" "2021-10-25" "2021-10-26"
#> [36] "2021-10-27" "2021-10-28" "2021-10-29" "2021-10-30" "2021-10-31"
#> [41] "2021-11-01" "2021-11-02" "2021-11-03" "2021-11-04" "2021-11-05"
#> [46] "2021-11-06" "2021-11-07" "2021-11-08" "2021-11-09" "2021-11-10"
#> [51] "2021-11-11" "2021-11-12" "2021-11-13" "2021-11-14" "2021-11-15"
#> [56] "2021-11-16" "2021-11-17" "2021-11-18" "2021-11-19" "2021-11-20"
#> [61] "2021-11-21" "2021-11-22" "2021-11-23" "2021-11-24" "2021-11-25"
#> [66] "2021-11-26" "2021-11-27" "2021-11-28" "2021-11-29" "2021-11-30"
#> [71] "2021-12-01" "2021-12-02" "2021-12-03" "2021-12-04" "2021-12-05"
#> [76] "2021-12-06" "2021-12-07" "2021-12-08" "2021-12-09" "2021-12-10"
#> [81] "2021-12-11" "2021-12-12" "2021-12-13" "2021-12-14" "2021-12-15"
#> [86] "2021-12-16" "2021-12-17" "2021-12-18" "2021-12-19" "2021-12-20"
#> [91] "2021-12-21" "2021-12-22"
#> 
#> $orbit_dates$RGT_cycle_14
#>  [1] "2021-12-22" "2021-12-23" "2021-12-24" "2021-12-25" "2021-12-26"
#>  [6] "2021-12-27" "2021-12-28" "2021-12-29" "2021-12-30" "2021-12-31"
#> [11] "2022-01-01" "2022-01-02" "2022-01-03" "2022-01-04" "2022-01-05"
#> [16] "2022-01-06" "2022-01-07" "2022-01-08" "2022-01-09" "2022-01-10"
#> [21] "2022-01-11" "2022-01-12" "2022-01-13" "2022-01-14" "2022-01-15"
#> [26] "2022-01-16" "2022-01-17" "2022-01-18" "2022-01-19" "2022-01-20"
#> [31] "2022-01-21" "2022-01-22" "2022-01-23" "2022-01-24" "2022-01-25"
#> [36] "2022-01-26" "2022-01-27" "2022-01-28" "2022-01-29" "2022-01-30"
#> [41] "2022-01-31" "2022-02-01" "2022-02-02" "2022-02-03" "2022-02-04"
#> [46] "2022-02-05" "2022-02-06" "2022-02-07" "2022-02-08" "2022-02-09"
#> [51] "2022-02-10" "2022-02-11" "2022-02-12" "2022-02-13" "2022-02-14"
#> [56] "2022-02-15" "2022-02-16" "2022-02-17" "2022-02-18" "2022-02-19"
#> [61] "2022-02-20" "2022-02-21" "2022-02-22" "2022-02-23" "2022-02-24"
#> [66] "2022-02-25" "2022-02-26" "2022-02-27" "2022-02-28" "2022-03-01"
#> [71] "2022-03-02" "2022-03-03" "2022-03-04" "2022-03-05" "2022-03-06"
#> [76] "2022-03-07" "2022-03-08" "2022-03-09" "2022-03-10" "2022-03-11"
#> [81] "2022-03-12" "2022-03-13" "2022-03-14" "2022-03-15" "2022-03-16"
#> [86] "2022-03-17" "2022-03-18" "2022-03-19" "2022-03-20" "2022-03-21"
#> [91] "2022-03-22" "2022-03-23"
#> 
#> $orbit_dates$RGT_cycle_15
#>  [1] "2022-03-23" "2022-03-24" "2022-03-25" "2022-03-26" "2022-03-27"
#>  [6] "2022-03-28" "2022-03-29" "2022-03-30" "2022-03-31" "2022-04-01"
#> [11] "2022-04-02" "2022-04-03" "2022-04-04" "2022-04-05" "2022-04-06"
#> [16] "2022-04-07" "2022-04-08" "2022-04-09" "2022-04-10" "2022-04-11"
#> [21] "2022-04-12" "2022-04-13" "2022-04-14" "2022-04-15" "2022-04-16"
#> [26] "2022-04-17" "2022-04-18" "2022-04-19" "2022-04-20" "2022-04-21"
#> [31] "2022-04-22" "2022-04-23" "2022-04-24" "2022-04-25" "2022-04-26"
#> [36] "2022-04-27" "2022-04-28" "2022-04-29" "2022-04-30" "2022-05-01"
#> [41] "2022-05-02" "2022-05-03" "2022-05-04" "2022-05-05" "2022-05-06"
#> [46] "2022-05-07" "2022-05-08" "2022-05-09" "2022-05-10" "2022-05-11"
#> [51] "2022-05-12" "2022-05-13" "2022-05-14" "2022-05-15" "2022-05-16"
#> [56] "2022-05-17" "2022-05-18" "2022-05-19" "2022-05-20" "2022-05-21"
#> [61] "2022-05-22" "2022-05-23" "2022-05-24" "2022-05-25" "2022-05-26"
#> [66] "2022-05-27" "2022-05-28" "2022-05-29" "2022-05-30" "2022-05-31"
#> [71] "2022-06-01" "2022-06-02" "2022-06-03" "2022-06-04" "2022-06-05"
#> [76] "2022-06-06" "2022-06-07" "2022-06-08" "2022-06-09" "2022-06-10"
#> [81] "2022-06-11" "2022-06-12" "2022-06-13" "2022-06-14" "2022-06-15"
#> [86] "2022-06-16" "2022-06-17" "2022-06-18" "2022-06-19" "2022-06-20"
#> [91] "2022-06-21"
#> 
#> $orbit_dates$RGT_cycle_16
#>  [1] "2022-06-21" "2022-06-22" "2022-06-23" "2022-06-24" "2022-06-25"
#>  [6] "2022-06-26" "2022-06-27" "2022-06-28" "2022-06-29" "2022-06-30"
#> [11] "2022-07-01" "2022-07-02" "2022-07-03" "2022-07-04" "2022-07-05"
#> [16] "2022-07-06" "2022-07-07" "2022-07-08" "2022-07-09" "2022-07-10"
#> [21] "2022-07-11" "2022-07-12" "2022-07-13" "2022-07-14" "2022-07-15"
#> [26] "2022-07-16" "2022-07-17" "2022-07-18" "2022-07-19" "2022-07-20"
#> [31] "2022-07-21" "2022-07-22" "2022-07-23" "2022-07-24" "2022-07-25"
#> [36] "2022-07-26" "2022-07-27" "2022-07-28" "2022-07-29" "2022-07-30"
#> [41] "2022-07-31" "2022-08-01" "2022-08-02" "2022-08-03" "2022-08-04"
#> [46] "2022-08-05" "2022-08-06" "2022-08-07" "2022-08-08" "2022-08-09"
#> [51] "2022-08-10" "2022-08-11" "2022-08-12" "2022-08-13" "2022-08-14"
#> [56] "2022-08-15" "2022-08-16" "2022-08-17" "2022-08-18" "2022-08-19"
#> [61] "2022-08-20" "2022-08-21" "2022-08-22" "2022-08-23" "2022-08-24"
#> [66] "2022-08-25" "2022-08-26" "2022-08-27" "2022-08-28" "2022-08-29"
#> [71] "2022-08-30" "2022-08-31" "2022-09-01" "2022-09-02" "2022-09-03"
#> [76] "2022-09-04" "2022-09-05" "2022-09-06" "2022-09-07" "2022-09-08"
#> [81] "2022-09-09" "2022-09-10" "2022-09-11" "2022-09-12" "2022-09-13"
#> [86] "2022-09-14" "2022-09-15" "2022-09-16" "2022-09-17" "2022-09-18"
#> [91] "2022-09-19" "2022-09-20"
#> 
#> $orbit_dates$RGT_cycle_17
#>  [1] "2022-09-20" "2022-09-21" "2022-09-22" "2022-09-23" "2022-09-24"
#>  [6] "2022-09-25" "2022-09-26" "2022-09-27" "2022-09-28" "2022-09-29"
#> [11] "2022-09-30" "2022-10-01" "2022-10-02" "2022-10-03" "2022-10-04"
#> [16] "2022-10-05" "2022-10-06" "2022-10-07" "2022-10-08" "2022-10-09"
#> [21] "2022-10-10" "2022-10-11" "2022-10-12" "2022-10-13" "2022-10-14"
#> [26] "2022-10-15" "2022-10-16" "2022-10-17" "2022-10-18" "2022-10-19"
#> [31] "2022-10-20" "2022-10-21" "2022-10-22" "2022-10-23" "2022-10-24"
#> [36] "2022-10-25" "2022-10-26" "2022-10-27" "2022-10-28" "2022-10-29"
#> [41] "2022-10-30" "2022-10-31" "2022-11-01" "2022-11-02" "2022-11-03"
#> [46] "2022-11-04" "2022-11-05" "2022-11-06" "2022-11-07" "2022-11-08"
#> [51] "2022-11-09" "2022-11-10" "2022-11-11" "2022-11-12" "2022-11-13"
#> [56] "2022-11-14" "2022-11-15" "2022-11-16" "2022-11-17" "2022-11-18"
#> [61] "2022-11-19" "2022-11-20" "2022-11-21" "2022-11-22" "2022-11-23"
#> [66] "2022-11-24" "2022-11-25" "2022-11-26" "2022-11-27" "2022-11-28"
#> [71] "2022-11-29" "2022-11-30" "2022-12-01" "2022-12-02" "2022-12-03"
#> [76] "2022-12-04" "2022-12-05" "2022-12-06" "2022-12-07" "2022-12-08"
#> [81] "2022-12-09" "2022-12-10" "2022-12-11" "2022-12-12" "2022-12-13"
#> [86] "2022-12-14" "2022-12-15" "2022-12-16" "2022-12-17" "2022-12-18"
#> [91] "2022-12-19" "2022-12-20"
#> 
#> $orbit_dates$RGT_cycle_18
#>  [1] "2022-12-20" "2022-12-21" "2022-12-22" "2022-12-23" "2022-12-24"
#>  [6] "2022-12-25" "2022-12-26" "2022-12-27" "2022-12-28" "2022-12-29"
#> [11] "2022-12-30" "2022-12-31" "2023-01-01" "2023-01-02" "2023-01-03"
#> [16] "2023-01-04" "2023-01-05" "2023-01-06" "2023-01-07" "2023-01-08"
#> [21] "2023-01-09" "2023-01-10" "2023-01-11" "2023-01-12" "2023-01-13"
#> [26] "2023-01-14" "2023-01-15" "2023-01-16" "2023-01-17" "2023-01-18"
#> [31] "2023-01-19" "2023-01-20" "2023-01-21" "2023-01-22" "2023-01-23"
#> [36] "2023-01-24" "2023-01-25" "2023-01-26" "2023-01-27" "2023-01-28"
#> [41] "2023-01-29" "2023-01-30" "2023-01-31" "2023-02-01" "2023-02-02"
#> [46] "2023-02-03" "2023-02-04" "2023-02-05" "2023-02-06" "2023-02-07"
#> [51] "2023-02-08" "2023-02-09" "2023-02-10" "2023-02-11" "2023-02-12"
#> [56] "2023-02-13" "2023-02-14" "2023-02-15" "2023-02-16" "2023-02-17"
#> [61] "2023-02-18" "2023-02-19" "2023-02-20" "2023-02-21" "2023-02-22"
#> [66] "2023-02-23" "2023-02-24" "2023-02-25" "2023-02-26" "2023-02-27"
#> [71] "2023-02-28" "2023-03-01" "2023-03-02" "2023-03-03" "2023-03-04"
#> [76] "2023-03-05" "2023-03-06" "2023-03-07" "2023-03-08" "2023-03-09"
#> [81] "2023-03-10" "2023-03-11" "2023-03-12" "2023-03-13" "2023-03-14"
#> [86] "2023-03-15" "2023-03-16" "2023-03-17" "2023-03-18" "2023-03-19"
#> [91] "2023-03-20" "2023-03-21"
#> 
#> $orbit_dates$RGT_cycle_19
#>  [1] "2023-03-21" "2023-03-22" "2023-03-23" "2023-03-24" "2023-03-25"
#>  [6] "2023-03-26" "2023-03-27" "2023-03-28" "2023-03-29" "2023-03-30"
#> [11] "2023-03-31" "2023-04-01" "2023-04-02" "2023-04-03" "2023-04-04"
#> [16] "2023-04-05" "2023-04-06" "2023-04-07" "2023-04-08" "2023-04-09"
#> [21] "2023-04-10" "2023-04-11" "2023-04-12" "2023-04-13" "2023-04-14"
#> [26] "2023-04-15" "2023-04-16" "2023-04-17" "2023-04-18" "2023-04-19"
#> [31] "2023-04-20" "2023-04-21" "2023-04-22" "2023-04-23" "2023-04-24"
#> [36] "2023-04-25" "2023-04-26" "2023-04-27" "2023-04-28" "2023-04-29"
#> [41] "2023-04-30" "2023-05-01" "2023-05-02" "2023-05-03" "2023-05-04"
#> [46] "2023-05-05" "2023-05-06" "2023-05-07" "2023-05-08" "2023-05-09"
#> [51] "2023-05-10" "2023-05-11" "2023-05-12" "2023-05-13" "2023-05-14"
#> [56] "2023-05-15" "2023-05-16" "2023-05-17" "2023-05-18" "2023-05-19"
#> [61] "2023-05-20" "2023-05-21" "2023-05-22" "2023-05-23" "2023-05-24"
#> [66] "2023-05-25" "2023-05-26" "2023-05-27" "2023-05-28" "2023-05-29"
#> [71] "2023-05-30" "2023-05-31" "2023-06-01" "2023-06-02" "2023-06-03"
#> [76] "2023-06-04" "2023-06-05" "2023-06-06" "2023-06-07" "2023-06-08"
#> [81] "2023-06-09" "2023-06-10" "2023-06-11" "2023-06-12" "2023-06-13"
#> [86] "2023-06-14" "2023-06-15" "2023-06-16" "2023-06-17" "2023-06-18"
#> [91] "2023-06-19" "2023-06-20"
#> 
#> $orbit_dates$RGT_cycle_20
#>  [1] "2023-06-20" "2023-06-21" "2023-06-22" "2023-06-23" "2023-06-24"
#>  [6] "2023-06-25" "2023-06-26" "2023-06-27" "2023-06-28" "2023-06-29"
#> [11] "2023-06-30" "2023-07-01" "2023-07-02" "2023-07-03" "2023-07-04"
#> [16] "2023-07-05" "2023-07-06" "2023-07-07" "2023-07-08" "2023-07-09"
#> [21] "2023-07-10" "2023-07-11" "2023-07-12" "2023-07-13" "2023-07-14"
#> [26] "2023-07-15" "2023-07-16" "2023-07-17" "2023-07-18" "2023-07-19"
#> [31] "2023-07-20" "2023-07-21" "2023-07-22" "2023-07-23" "2023-07-24"
#> [36] "2023-07-25" "2023-07-26" "2023-07-27" "2023-07-28" "2023-07-29"
#> [41] "2023-07-30" "2023-07-31" "2023-08-01" "2023-08-02" "2023-08-03"
#> [46] "2023-08-04" "2023-08-05" "2023-08-06" "2023-08-07" "2023-08-08"
#> [51] "2023-08-09" "2023-08-10" "2023-08-11" "2023-08-12" "2023-08-13"
#> [56] "2023-08-14" "2023-08-15" "2023-08-16" "2023-08-17" "2023-08-18"
#> [61] "2023-08-19" "2023-08-20" "2023-08-21" "2023-08-22" "2023-08-23"
#> [66] "2023-08-24" "2023-08-25" "2023-08-26" "2023-08-27" "2023-08-28"
#> [71] "2023-08-29" "2023-08-30" "2023-08-31" "2023-09-01" "2023-09-02"
#> [76] "2023-09-03" "2023-09-04" "2023-09-05" "2023-09-06" "2023-09-07"
#> [81] "2023-09-08" "2023-09-09" "2023-09-10" "2023-09-11" "2023-09-12"
#> [86] "2023-09-13" "2023-09-14" "2023-09-15" "2023-09-16" "2023-09-17"
#> [91] "2023-09-18" "2023-09-19"
#> 
#> 
#> $length_dates
#> $length_dates$RGT_cycle_1
#> [1] 77
#> 
#> $length_dates$RGT_cycle_2
#> [1] 92
#> 
#> $length_dates$RGT_cycle_3
#> [1] 92
#> 
#> $length_dates$RGT_cycle_4
#> [1] 80
#> 
#> $length_dates$RGT_cycle_5
#> [1] 92
#> 
#> $length_dates$RGT_cycle_6
#> [1] 92
#> 
#> $length_dates$RGT_cycle_7
#> [1] 92
#> 
#> $length_dates$RGT_cycle_8
#> [1] 92
#> 
#> $length_dates$RGT_cycle_9
#> [1] 91
#> 
#> $length_dates$RGT_cycle_10
#> [1] 91
#> 
#> $length_dates$RGT_cycle_11
#> [1] 92
#> 
#> $length_dates$RGT_cycle_12
#> [1] 92
#> 
#> $length_dates$RGT_cycle_13
#> [1] 92
#> 
#> $length_dates$RGT_cycle_14
#> [1] 92
#> 
#> $length_dates$RGT_cycle_15
#> [1] 91
#> 
#> $length_dates$RGT_cycle_16
#> [1] 92
#> 
#> $length_dates$RGT_cycle_17
#> [1] 92
#> 
#> $length_dates$RGT_cycle_18
#> [1] 92
#> 
#> $length_dates$RGT_cycle_19
#> [1] 92
#> 
#> $length_dates$RGT_cycle_20
#> [1] 92
#> 
#> 

#...................................................
# observe and choose one of the available RGT-cycles
#...................................................

avail_cycles = available_RGTs(only_cycle_names = TRUE,
                              verbose = TRUE)
#> The available Icesat-2 orbits will be red from 'https://icesat-2.gsfc.nasa.gov/science/specs' ... 
#> Access the data of the technical specs website ...
#> Extract the .zip files and the corresponding titles ...
#> Keep the relevant data from the url's and titles ...
#> Process the nominal and time specific orbits separately ...
#> Adjust the Dates of the time specific orbits ...
#> Create the nominal orbits data.table ...
#> Create the time specific orbits data.table ...
#> Return a single data.table ...
#> Elapsed time: 0 hours and 0 minutes and 0 seconds. 
avail_cycles
#>  [1] "RGT_cycle_1"  "RGT_cycle_2"  "RGT_cycle_3"  "RGT_cycle_4"  "RGT_cycle_5" 
#>  [6] "RGT_cycle_6"  "RGT_cycle_7"  "RGT_cycle_8"  "RGT_cycle_9"  "RGT_cycle_10"
#> [11] "RGT_cycle_11" "RGT_cycle_12" "RGT_cycle_13" "RGT_cycle_14" "RGT_cycle_15"
#> [16] "RGT_cycle_16" "RGT_cycle_17" "RGT_cycle_18" "RGT_cycle_19" "RGT_cycle_20"

#.....................................
# receive results for a specific cycle
#.....................................

rev_cycle = revisit_time_RGTs(RGT_cycle = 'RGT_cycle_1', complete_date_sequence = FALSE)
rev_cycle
#> $orbit_files
#> [1] "https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle1_date_time.zip"
#> 
#> $orbit_dates
#> [1] "2018-10-13" "2018-12-28"
#> 
#> $length_dates
#> [1] 77
#>